Add conditional rendering for download QR button in Pet page
This commit adds a conditional rendering logic for the download QR button in the Pet page. The button will only be displayed if the logged-in user is the owner of the pet. This improves the user experience by showing the button only to relevant users and avoids confusion for other users.