feat: Add ConfirmationDialog component
This commit adds a new component called ConfirmationDialog to the atoms folder. The component is exported from the index.ts file. This component will be used to display confirmation dialogs throughout the application.

chore: Add LiveLocation store

This commit adds a new store called LiveLocation to the lib/store folder. The store manages the state related to live location tracking. It includes properties for tracking status, latitude, and longitude, as well as functions to start and stop tracking and set the live location.

chore: Update CustomInput component

This commit updates the CustomInput component in the molecules folder. It adds a new optional prop called type, which allows specifying the input type. If no type is provided, it defaults to "text".

chore: Update App component

This commit updates the App component in the App.tsx file. It adds functionality to track the user's live location when the tracking flag is enabled and there is a logged-in user. The user's location is obtained using the geolocation API and sent to the server using the setUsersLocation function from the users.service module.

chore: Update AppBarMenu component

This commit updates the AppBarMenu component in the atoms folder. It adds a new menu item for editing the user's profile. The item includes an icon and text, and when clicked, it navigates to the profile edit page.

chore: Update users.service module

This commit updates the users.service module in the lib/services folder. It adds a new function called updateUser, which allows updating the user's name, lastname, date of birth, and phone number. It also adds a new function called setUsersLocation, which allows setting the user's location coordinates.