feat: Add copy functionality for email and phone in UserDetails
This commit adds the functionality to copy the email and phone number of a user in the UserDetails component. It includes the handleCopyEmail and handleCopyPhone functions that use the navigator.clipboard API to write the email and phone number to the clipboard. Toast messages are displayed to indicate successful copying.