feat: Add geolocation tracking and set user's location
This commit adds geolocation tracking functionality to the `startTracking` method in the `useLiveLocationStore` hook. When the method is called, it checks if the browser supports geolocation. If it does, it retrieves the current position and sets the user's location using the `setUsersLocation` function from the `users.service` module. The latitude and longitude are then stored in the state, and the tracking flag is set to true. If geolocation is not supported, an error message is logged to the console.